@@ -13,43 +13,42 @@
   "dependencies": {
     "classnames": "^2.2.6",
     "dotenv": "^8.2.0",
-    "gatsby": "^2.24.63",
+    "gatsby": "^2.24.65",
     "gatsby-image": "^2.4.19",
     "gatsby-link": "^2.4.14",
     "gatsby-plugin-advanced-sitemap": "^1.5.6",
-    "gatsby-plugin-manifest": "^2.4.30",
-    "gatsby-plugin-preact": "^3.1.26",
+    "gatsby-plugin-manifest": "^2.4.31",
     "gatsby-plugin-react-helmet": "^3.3.11",
     "gatsby-plugin-react-svg": "^3.0.0",
     "gatsby-plugin-sharp": "^2.6.36",
     "gatsby-plugin-styled-jsx": "^3.3.12",
     "gatsby-source-filesystem": "^2.3.30",
     "gatsby-source-wordpress": "^3.3.34",
+    "gatsby-source-wordpress-experimental": "^1.4.0",
     "gatsby-transformer-sharp": "^2.5.15",
-    "http-proxy-middleware": "^0.20.0",
-    "preact": "^10.2.1",
-    "rc-slider": "^9.4.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-helmet": "^5.2.1",
-    "react-select": "^3.0.5",
-    "styled-jsx": "^3.2.4",
+    "http-proxy-middleware": "^1.0.5",
+    "rc-slider": "^9.5.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-helmet": "^6.1.0",
+    "react-select": "^3.1.0",
+    "styled-jsx": "^3.3.0",
     "twit": "^2.2.11",
-    "twitter-text": "^3.0.1"
+    "twitter-text": "^3.1.0"
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-react": "^7.18.3",
-    "husky": "^4.2.1",
-    "lint-staged": "^10.4.0",
-    "netlify-lambda": "^1.6.3",
-    "prettier": "^1.19.1"
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.10.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-prettier": "^6.12.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-react": "^7.21.2",
+    "husky": "^4.3.0",
+    "lint-staged": "^10.3.0",
+    "netlify-lambda": "^2.0.1",
+    "prettier": "^2.1.1"
   },
   "husky": {
     "hooks": {
