@@ -1,9 +1,8 @@
 import { graphql } from 'gatsby';
-import Helmet from 'react-helmet';
+import { Helmet } from 'react-helmet';
 import React from 'react';
 
 import Constraint from '../../components/constraint';
-import Intro from '../../components/intro';
 import Richtext from '../../components/richtext';
 import style from './style';
 import withLayout from '../../components/with-layout';
@@ -12,7 +11,7 @@ const Page = ({
   data: {
     page: {
       title,
-      acf: { content, intro },
+      acf: { content },
     },
   },
 }) => (
@@ -25,15 +24,13 @@ const Page = ({
       <Constraint topLevel>
         <h1 dangerouslySetInnerHTML={{ __html: title }} />
 
-        <Intro intro={intro} />
-
         {content &&
           content.map((block) => {
             const { __typename: type } = block;
 
             // eslint-disable-next-line default-case
             switch (type) {
-              case 'WordPressAcf_text':
+              case 'WpPage_Acf_Content_Text':
                 return <Richtext content={block.text} />;
             }
 
@@ -48,13 +45,13 @@ export default withLayout(Page);
 
 export const query = graphql`
   query($wordpressId: Int) {
-    page: wordpressPage(wordpress_id: { eq: $wordpressId }) {
+    page: wpPage(databaseId: { eq: $wordpressId }) {
       title
       acf {
-        content: content_page {
+        content {
           __typename
 
-          ... on WordPressAcf_text {
+          ... on WpPage_Acf_Content_Text {
             text
           }
         }
