@@ -1,5 +1,5 @@
 import { graphql } from 'gatsby';
-import Helmet from 'react-helmet';
+import { Helmet } from 'react-helmet';
 import React from 'react';
 
 import AuthorList from '../../components/author-list';
@@ -28,11 +28,11 @@ const Page = ({
         language,
         institute,
         employer,
-        published_in: publishedIn,
+        publishedIn,
       },
       title,
       featuredImage,
-      tags,
+      tags: { nodes: tags },
     },
   },
 }) => (
@@ -63,26 +63,24 @@ const Page = ({
             <div className="languages-container">
               {buttonIcon.styles}
 
-              {language.map(
-                ({ language: downloadLanguage, external_url: externalUrl }) => (
-                  <Button
-                    key={`language-${downloadLanguage}`}
-                    to={externalUrl}
-                    theme="blue"
-                    external
-                  >
-                    {HUMAN_READABLE_LANGUAGES[downloadLanguage]}
-                    <DownloadIcon className={buttonIcon.className} />
-                  </Button>
-                )
-              )}
+              {language.map(({ language: downloadLanguage, externalUrl }) => (
+                <Button
+                  key={`language-${downloadLanguage}`}
+                  to={externalUrl}
+                  theme="blue"
+                  external
+                >
+                  {HUMAN_READABLE_LANGUAGES[downloadLanguage]}
+                  <DownloadIcon className={buttonIcon.className} />
+                </Button>
+              ))}
             </div>
           )}
         </div>
 
-        {featuredImage && featuredImage.localFile && (
+        {featuredImage?.node?.localFile && (
           <div className="cover-image-container">
-            <Picture image={featuredImage.localFile} />
+            <Picture image={featuredImage.node.localFile} />
           </div>
         )}
       </header>
@@ -144,27 +142,29 @@ export default withLayout(Page);
 
 export const query = graphql`
   query($wordpressId: Int) {
-    publication: wordpressWpPublications(wordpress_id: { eq: $wordpressId }) {
+    publication: wpPublication(databaseId: { eq: $wordpressId }) {
       title
-      featuredImage: featured_media {
-        localFile {
-          childImageSharp {
-            fluid(maxWidth: 800) {
-              src
-              srcSet
-              srcSetWebp
+      featuredImage {
+        node {
+          localFile {
+            childImageSharp {
+              fluid(maxWidth: 800) {
+                ...Picture
+              }
             }
           }
         }
       }
       tags {
-        name
-        slug
+        nodes {
+          name
+          slug
+        }
       }
       acf {
         abstract
         year
-        published_in
+        publishedIn
         subtitle
         author {
           name
@@ -175,7 +175,7 @@ export const query = graphql`
         }
         language {
           language
-          external_url
+          externalUrl
           file {
             link
           }
